+++ api/dbmodels/phone/mock_phone/querier_mock.go
@@ -35,7 +35,7 @@ func (m *MockQuerier) EXPECT() *MockQuerierMockRecorder {
 }
 
 // GetByID mocks base method
-func (m *MockQuerier) GetByID(ctx context.Context, id int32) (phone.Phone, error) {
+func (m *MockQuerier) GetByID(ctx context.Context, id int64) (phone.Phone, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetByID", ctx, id)
 	ret0, _ := ret[0].(phone.Phone)
